// frontend/src/shared/api/model/paginatedCourseEnrollmentList.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Student Evaluation System API
 * Outcome-based assessment system for academic programs
 * OpenAPI spec version: 1.0.0
 */
import type { CourseEnrollment } from './courseEnrollment';

export interface PaginatedCourseEnrollmentList {
  count: number;
  /** @nullable */
  next?: string | null;
  /** @nullable */
  previous?: string | null;
  results: CourseEnrollment[];
}